update firestore class
