--- exercices/firestoreHelper.dart
@@ -1,5 +1,0 @@
-class FirestoreHelper{
-
-// add users
-addUser(uid ,map)
-
